Login adicionado, com token authentication. E alguns routers permitidos apenas para admins

+++ endpoints/admins.py
@@ -2,6 +2,8 @@ import models, schemas
 from database import SessionLocal
 from sqlalchemy.orm import Session
 from fastapi import Depends, APIRouter
+from models import *
+from endpoints.loginn import *
 
 router = APIRouter()
 
@@ -13,16 +15,26 @@ def get_session(): #Função para pegar a sessão, e abrir e fechar o banco de d
         session.close()
 
 @router.get("/admins")
-def getItems(session: Session = Depends(get_session)): #Pegando os valores do banco de dados, Depends do get_session
+def getItems(session: Session = Depends(get_session), user: Cadastro_Users = Depends(get_current_user)): #Pegando os valores do banco de dados, Depends do get_session
+    if not user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Você não tem permissão para acessar os admins!"
+        )
     items = session.query(models.Cadastro_Admins).all()
-    return items
+    return {f"Olá {user.username}, aqui estão os admins: {items}"}
 
 #Aqui a gente chama a classe responsável pelos valores que vão ser necessitados aqui, e chamamos eles para passarem os valores e serem encaminhados para o banco de dados
 @router.get("/admins/{id}") #Router para trazer as informações de acordo com o id
-def getItem(id:int, session: Session = Depends(get_session)): #Criando um getItem, (get). que espera receber uma variável (id) e com os dois pontos :int eu EXIJO que a variável que venha seja INT
+def getItem(id:int, session: Session = Depends(get_session), user: Cadastro_Users = Depends(get_current_user)): #Criando um getItem, (get). que espera receber uma variável (id) e com os dois pontos :int eu EXIJO que a variável que venha seja INT
+    if not user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Você não tem permissão para acessar os admins!"
+        )
     item = session.query(models.Cadastro_Admins).get(id) #Para pegar apenas o valor que for representado pelo id
     #return fakeDataBase[id] #Retornando o valor do dicionário, de acordo com o ID que ele digitar
-    return item
+    return {f"Olá {user.username}, aqui está o admin id:{id} - {item}"}
 
 ###################################################################################################################################################
  #1 tipo de fazer um post, mas apenas é recomendado se for passar 1 ou poucos valores como é o exemplo... caso contrário é melhor fazer de outra forma como o pydantic 
@@ -36,7 +48,12 @@ def getItem(id:int, session: Session = Depends(get_session)): #Criando um getIte
 #2 Segundo jeito de fazer e o melhor para vários valores, seria com o pydantic, para passar mais valores e informações, como strings, ints e etc
 #Aqui a gente chama a classe responsável pelos valores que vão ser necessitados aqui, e chamamos eles para passarem os valores e serem encaminhados para o banco de dados
 @router.post("/admins/adicionar")
-def addItem(item:schemas.Admins, session: Session = Depends(get_session)): #Aqui se chamaria a classe, e o nome do classe dentro da classe, para pegar os valores e fazer um objeto
+def addItem(item:schemas.Admins, session: Session = Depends(get_session), user: Cadastro_Users = Depends(get_current_user)): #Aqui se chamaria a classe, e o nome do classe dentro da classe, para pegar os valores e fazer um objeto
+    if not user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Você não tem permissão para criar um admin!"
+        )
     #newId = len(fakeDataBase.keys()) + 1 #Pegando o tamanho do fakedatabase e adicionando o valor de +1 para que o próximo item que for adicionado seja na próxima key
     item = models.Cadastro_Admins(nome = item.nome, email = item.email, senha = item.senha)
     session.add(item) #Adicionando no banco
@@ -44,7 +61,7 @@ def addItem(item:schemas.Admins, session: Session = Depends(get_session)): #Aqui
     session.refresh(item) #Dando um refresh para atualizar o banco
     #fakeDataBase[newId] = {"task":item.task, "valor":item.value} #Indicando que no fakedatabase no novo id, seja adicionado task + o valor que for passado
     #return fakeDataBase
-    return item
+    return {f"Olá {user.username}, adicionando o admin: {item}"}
 
 ####################################################################################################################################################
 #3 Seria um jeito criando uma requisição com o próprio Body (mas tem muitos problemas desse jeito e não é mto recomendado)
@@ -56,23 +73,32 @@ def addItem(item:schemas.Admins, session: Session = Depends(get_session)): #Aqui
 
 #Atualizando valores
 @router.put("/admins/atualizar/{id}")
-def updateItem(id:int, item:schemas.Admins, session: Session = Depends(get_session)): #Aqui se chamaria a classe, e o nome do classe dentro da classe, para pegar os valores e fazer um objeto
+def updateItem(id:int, item:schemas.Admins, session: Session = Depends(get_session), user: Cadastro_Users = Depends(get_current_user)): #Aqui se chamaria a classe, e o nome do classe dentro da classe, para pegar os valores e fazer um objeto
+    if not user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Você não tem permissão de atualizar admins!"
+        )
     #newId = len(fakeDataBase.keys()) + 1 #Pegando o tamanho do fakedatabase e adicionando o valor de +1 para que o próximo item que for adicionado seja na próxima key
     itemObject = session.query(models.Cadastro_Admins).get(id) #Pegando o valor que foi passado pelo int, de qual objeto salvo é
     itemObject.nome, itemObject.email, itemObject.senha = item.nome, item.email, item.senha #atualizando com esses valores novos 
     session.commit() #comitando a mudança
     #fakeDataBase[newId] = {"task":item.task, "valor":item.value} #Indicando que no fakedatabase no novo id, seja adicionado task + o valor que for passado
     #return fakeDataBase
-    return itemObject
+    return {f"Olá {user.username}, atualizando o admin:{id} para - {itemObject}"}
 
 #Deletando valores
 @router.delete("/admins/delete/{id}")
-def deleteItem(id:int, session: Session = Depends(get_session)): #Aqui se chamaria a classe, e o nome do classe dentro da classe, para pegar os valores e fazer um objeto
+def deleteItem(id:int, session: Session = Depends(get_session), user: Cadastro_Users = Depends(get_current_user)): #Aqui se chamaria a classe, e o nome do classe dentro da classe, para pegar os valores e fazer um objeto
+    if not user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Você não tem permissão para excluir admins!"
+        )
     #newId = len(fakeDataBase.keys()) + 1 #Pegando o tamanho do fakedatabase e adicionando o valor de +1 para que o próximo item que for adicionado seja na próxima key
     itemObject = session.query(models.Cadastro_Admins).get(id) #Pegando o valor que foi passado pelo int, de qual objeto salvo é
     session.delete(itemObject) #comitando a mudança
     session.commit() #Comitando as mudanças
     session.close() #Fechando o banco
-    mensagem = 'O item {} foi deletado'.format(id)
-    return  mensagem, itemObject
+    return {f"Olá {user.username}, deletando o admin de id: {id} - {itemObject}"}
